Move more files into automated pipelines directory (#37415)
