+++ src/automated-pipelines/components/AutomatedPage.tsx
@@ -3,9 +3,9 @@ import { ArticleTitle } from 'components/article/ArticleTitle'
 import { MarkdownContent } from 'components/ui/MarkdownContent'
 import { Lead } from 'components/ui/Lead'
 import { PermissionsStatement } from 'components/ui/PermissionsStatement'
-import { ArticleGridLayout } from './ArticleGridLayout'
+import { ArticleGridLayout } from '../../../components/article/ArticleGridLayout'
 import { MiniTocs } from 'components/ui/MiniTocs'
-import { useAutomatedPageContext } from 'components/context/AutomatedPageContext'
+import { useAutomatedPageContext } from 'src/automated-pipelines/components/AutomatedPageContext'
 import { ClientSideHighlight } from 'components/ClientSideHighlight'
 import { Callout } from 'components/ui/Callout'
