+++ src/automated-pipelines/components/AutomatedPageContext.tsx
