+++ src/automated-pipelines/components/parameter-table/ChildBodyParametersRows.tsx
