+++ src/automated-pipelines/components/parameter-table/ParameterRow.tsx
