+++ src/automated-pipelines/components/parameter-table/ParameterTable.module.scss
