+++ src/automated-pipelines/components/parameter-table/ParameterTable.tsx
