+++ src/automated-pipelines/components/parameter-table/types.ts
