+++ src/graphql/pages/breaking-changes.tsx
@@ -2,12 +2,12 @@ import { GetServerSideProps } from 'next'
 import React from 'react'
 
 import { MainContextT, MainContext, getMainContext } from 'components/context/MainContext'
-import { AutomatedPage } from 'components/article/AutomatedPage'
+import { AutomatedPage } from 'src/automated-pipelines/components/AutomatedPage'
 import {
   AutomatedPageContext,
   AutomatedPageContextT,
   getAutomatedPageContextFromRequest,
-} from 'components/context/AutomatedPageContext'
+} from 'src/automated-pipelines/components/AutomatedPageContext'
 import { BreakingChanges } from 'src/graphql/components/BreakingChanges'
 import { BreakingChangesT } from 'src/graphql/components/types'
