+++ src/graphql/pages/schema-previews.tsx
@@ -2,12 +2,12 @@ import { GetServerSideProps } from 'next'
 import React from 'react'
 
 import { MainContextT, MainContext, getMainContext } from 'components/context/MainContext'
-import { AutomatedPage } from 'components/article/AutomatedPage'
+import { AutomatedPage } from 'src/automated-pipelines/components/AutomatedPage'
 import {
   AutomatedPageContext,
   AutomatedPageContextT,
   getAutomatedPageContextFromRequest,
-} from 'components/context/AutomatedPageContext'
+} from 'src/automated-pipelines/components/AutomatedPageContext'
 import { Previews } from 'src/graphql/components/Previews'
 import { PreviewT } from 'src/graphql/components/types'
