+++ src/rest/components/RestCollapsibleSection.tsx
@@ -5,7 +5,7 @@ import { TreeView } from '@primer/react'
 
 import { ProductTreeNode } from 'components/context/MainContext'
 import { EventType, sendEvent } from 'src/events/components/events'
-import { useAutomatedPageContext } from 'components/context/AutomatedPageContext'
+import { useAutomatedPageContext } from 'src/automated-pipelines/components/AutomatedPageContext'
 import type { MiniTocItem } from 'components/context/ArticleContext'
 import { Link } from 'components/Link'
