+++ src/rest/components/RestOperation.tsx
@@ -7,7 +7,7 @@ import { HeadingLink } from 'components/article/HeadingLink'
 import { Link } from 'components/Link'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { RestPreviewNotice } from './RestPreviewNotice'
-import { ParameterTable } from 'components/parameter-table/ParameterTable'
+import { ParameterTable } from 'src/automated-pipelines/components/parameter-table/ParameterTable'
 import { RestCodeSamples } from './RestCodeSamples'
 import { RestStatusCodes } from './RestStatusCodes'
 import { Operation } from './types'
