+++ src/rest/components/RestReferencePage.tsx
@@ -5,7 +5,7 @@ import { MarkdownContent } from 'components/ui/MarkdownContent'
 import { Lead } from 'components/ui/Lead'
 import { PermissionsStatement } from 'components/ui/PermissionsStatement'
 import { RestOperation } from './RestOperation'
-import { useAutomatedPageContext } from 'components/context/AutomatedPageContext'
+import { useAutomatedPageContext } from 'src/automated-pipelines/components/AutomatedPageContext'
 import { Operation } from './types'
 import { ClientSideRedirects } from 'src/rest/components/ClientSideRedirects'
 import { RestRedirect } from 'src/rest/components/RestRedirect'
